packaging RENAME package from libyang2 to just libyang
