+++ setup.py
@@ -11,11 +11,11 @@ def read_file(fpath, encoding="utf-8"):
         return f.read().strip()
 
 
-setup(name='libyang2',
+setup(name='libyang',
       description="CFFI bindings to libyang",
       version='1.0',
       setup_requires=["cffi>=1.0.0"],
       install_requires=["cffi>=1.0.0"],
       cffi_modules=["cffi/build.py:ffibuilder"],
-      packages=['libyang2'],
+      packages=['libyang'],
 )
